refactor!: move authenticate_by and recover! to Account::Password
**Motivation**

`Account`'s `Authenticatable` concern was still hard-coded to
email/password (`authenticate_by`, `authenticate`,
`prevent_timing_atack`), and `Recoverable#recover!` literally reset a
password — both auth-method-specific despite `Account` no longer
storing the password digest directly. `Account` should hold only
generic identity/session concerns; each auth method should own its own
entry point.

**Overview**

- `authenticate_by` (plus its `find_by_identifier`/timing-attack-prevention
  helpers) moved from `Account` to `Account::Password`
- `recover!` moved from `Account` to `Account::Password`;
  `Account::Recovery` still owns the generic token lifecycle
- `Lockable` stays as-is on `Account`: brute-force lockout is generic
  infrastructure any credential-guessing method could reuse; only the
  decision to trigger it moved, naturally, since it now lives inside
  `Password.authenticate_by`

**Main Changes**

- app/models/aikotoba/account.rb (change): remove `authenticate_by`/`authenticate`/`prevent_timing_atack`/`recover!`; keep only generic `find_by_identifier`, `authentication_failed!`/`success!`
- app/models/aikotoba/account/password.rb (change): add `authenticate_by`/`prevent_timing_attack` class methods and `recover!`
- app/models/aikotoba/account/recovery.rb (change): delegate password reset to `Account::Password#recover!`, re-import validation errors onto `:password`
- app/controllers/aikotoba/sessions_controller.rb, app/controllers/aikotoba/api/sessions_controller.rb (change): call `Account::Password.authenticate_by` instead of `Account.authenticate_by`
- test/models/aikotoba/account/password_test.rb (add): move `authenticate_by`/`recover!` test coverage here
- test/models/aikotoba/account_test.rb, test/controllers/aikotoba/recoverable_test.rb (change): update call sites, trim moved tests
- README.md (change): update usage example for `Account::Password.authenticate_by`

**Breaking Changes (Important)**

- `Aikotoba::Account.authenticate_by` no longer exists; use `Aikotoba::Account::Password.authenticate_by` instead
- `Aikotoba::Account#recover!` no longer exists; use `Aikotoba::Account::Recovery.recover!` (unchanged) or `Aikotoba::Account::Password#recover!` directly
