+++ README.md
@@ -523,7 +523,7 @@ By running the following script, you can hash and store passwords.
 
 ```ruby
 Aikotoba::Account.create!(email: "sample@example.com", password: "password")
-Aikotoba::Account.authenticate_by(attributes: {email: "sample@example.com", password: "password"})
+Aikotoba::Account::Password.authenticate_by(attributes: {email: "sample@example.com", password: "password"})
 # => created account instance.
 ```
