+++ app/controllers/aikotoba/sessions_controller.rb
@@ -40,7 +40,7 @@ module Aikotoba
     end
 
     def authenticate_account(params)
-      Account.authenticate_by(attributes: params, target_type_name: aikotoba_authenticate_target)
+      Account::Password.authenticate_by(attributes: params, target_type_name: aikotoba_authenticate_target)
     end
 
     def after_sign_in_path
