+++ test/models/aikotoba/account_test.rb
@@ -16,22 +16,6 @@ class Aikotoba::AccountTest < ActiveSupport::TestCase
       account.password = "Password1!"
       assert_equal account.password, "Password1!"
     end
-
-    test "password_match? returns true for correct password" do
-      account = Aikotoba::Account.create!(
-        email: "user@example.com",
-        password: "Password1!"
-      )
-      assert account.password_match?("Password1!")
-    end
-
-    test "password_match? returns false for incorrect password" do
-      account = Aikotoba::Account.create!(
-        email: "user@example.com",
-        password: "Password1!"
-      )
-      refute account.password_match?("WrongPassword!")
-    end
   end
 
   class Registrable < ActiveSupport::TestCase
@@ -90,16 +74,6 @@ class Aikotoba::AccountTest < ActiveSupport::TestCase
       )
     end
 
-    test "authenticate returns account for correct password" do
-      result = @account.authenticate("Password1!")
-      assert_equal result, @account
-    end
-
-    test "authenticate returns nil for incorrect password" do
-      result = @account.authenticate("WrongPassword!")
-      assert_nil result
-    end
-
     test "authentication_failed! increments failed_attempts" do
       assert_difference "@account.failed_attempts" do
         @account.authentication_failed!
@@ -110,65 +84,6 @@ class Aikotoba::AccountTest < ActiveSupport::TestCase
       @account.update!(failed_attempts: 5)
       @account.authentication_success!
       assert_equal @account.failed_attempts, 0
-    end
-
-    test "authenticate_by returns account for valid credentials" do
-      account = Aikotoba::Account.authenticate_by(attributes: {
-        email: "user@example.com",
-        password: "Password1!"
-      })
-      assert_equal account, @account
-    end
-
-    test "authenticate_by returns nil for invalid email" do
-      account = Aikotoba::Account.authenticate_by(attributes: {
-        email: "nonexistent@example.com",
-        password: "Password1!"
-      })
-      assert_nil account
-    end
-
-    test "authenticate_by returns nil for invalid password" do
-      account = Aikotoba::Account.authenticate_by(attributes: {
-        email: "user@example.com",
-        password: "WrongPassword!"
-      })
-      assert_nil account
-    end
-
-    test "authenticate_by increments failed_attempts on invalid password" do
-      initial_attempts = @account.failed_attempts
-      Aikotoba::Account.authenticate_by(attributes: {
-        email: "user@example.com",
-        password: "WrongPassword!"
-      })
-      assert_equal @account.reload.failed_attempts, initial_attempts + 1
-    end
-
-    test "authenticate_by resets failed_attempts on successful authentication" do
-      @account.update!(failed_attempts: 5)
-      Aikotoba::Account.authenticate_by(attributes: {
-        email: "user@example.com",
-        password: "Password1!"
-      })
-      assert_equal @account.reload.failed_attempts, 0
-    end
-
-    test "authenticate_by locks account when max attempts exceeded" do
-      Aikotoba.lockable = true
-      account = Aikotoba::Account.create!(
-        email: "locktest@example.com",
-        password: "Password1!",
-        confirmed: true,
-        locked: false,
-        failed_attempts: 10
-      )
-      Aikotoba::Account.authenticate_by(attributes: {
-        email: "locktest@example.com",
-        password: "WrongPassword!"
-      })
-      assert account.reload.locked?
-      Aikotoba.lockable = false
     end
 
     test "authenticatable scope returns all accounts when features disabled" do
@@ -345,29 +260,6 @@ class Aikotoba::AccountTest < ActiveSupport::TestCase
       account.unlock!
       assert_not account.reload.locked?
       assert_equal account.failed_attempts, 0
-    end
-  end
-
-  class Recoverable < ActiveSupport::TestCase
-    test "recover! updates password_digest" do
-      account = Aikotoba::Account.create!(
-        email: "user@example.com",
-        password: "OldPassword1!"
-      )
-      old_digest = account.password_digest
-      account.recover!(new_password: "NewPassword1!")
-      assert_not_equal account.reload.password_digest, old_digest
-      assert account.password_match?("NewPassword1!")
-    end
-
-    test "recover! validates password length" do
-      account = Aikotoba::Account.create!(
-        email: "user@example.com",
-        password: "Password1!"
-      )
-      assert_raises(ActiveRecord::RecordInvalid) do
-        account.recover!(new_password: "short")
-      end
     end
   end
